requirements.txt
flask==2.0.3
werkzeug==2.0.3
python-dotenv==0.19.0
PyMuPDF==1.23.8
faiss-cpu==1.7.3
openai==0.27.0
python-magic==0.4.24 
numpy==1.23.5
pandas==1.5.3
nltk==3.7
